extract log packer
