+++ core/services/ocr2/plugins/ocr2keeper/evm21/abi.go
@@ -8,9 +8,7 @@ import (
 	"github.com/ethereum/go-ethereum/common/hexutil"
 	ocr2keepers "github.com/smartcontractkit/ocr2keepers/pkg"
 
-	"github.com/smartcontractkit/chainlink/v2/core/chains/evm/logpoller"
 	iregistry21 "github.com/smartcontractkit/chainlink/v2/core/gethwrappers/generated/i_keeper_registry_master_wrapper_2_1"
-	"github.com/smartcontractkit/chainlink/v2/core/gethwrappers/generated/i_log_automation"
 )
 
 // enum UpkeepFailureReason is defined by https://github.com/smartcontractkit/chainlink/blob/develop/contracts/src/v0.8/dev/automation/2_1/interfaces/AutomationRegistryInterface2_1.sol#L97
@@ -35,29 +33,11 @@ const (
 type UpkeepInfo = iregistry21.KeeperRegistryBase21UpkeepInfo
 
 type evmRegistryPackerV2_1 struct {
-	abi        abi.ABI
-	logDataABI abi.ABI
+	abi abi.ABI
 }
 
-func NewEvmRegistryPackerV2_1(abi abi.ABI, logDataABI abi.ABI) *evmRegistryPackerV2_1 {
-	return &evmRegistryPackerV2_1{abi: abi, logDataABI: logDataABI}
-}
-
-func (rp *evmRegistryPackerV2_1) PackLogData(log logpoller.Log) ([]byte, error) {
-	topics := [][32]byte{}
-	for _, topic := range log.GetTopics() {
-		topics = append(topics, topic)
-	}
-	return rp.logDataABI.Pack("checkLog", &i_log_automation.Log{
-		Index:       big.NewInt(log.LogIndex),
-		TxIndex:     big.NewInt(0), // TODO
-		TxHash:      log.TxHash,
-		BlockNumber: big.NewInt(log.BlockNumber),
-		BlockHash:   log.BlockHash,
-		Source:      log.Address,
-		Topics:      topics,
-		Data:        log.Data,
-	})
+func NewEvmRegistryPackerV2_1(abi abi.ABI) *evmRegistryPackerV2_1 {
+	return &evmRegistryPackerV2_1{abi: abi}
 }
 
 // TODO: remove for 2.1
