+++ core/services/ocr2/plugins/ocr2keeper/evm21/feed_lookup_test.go
@@ -62,7 +62,7 @@ func setupEVMRegistry(t *testing.T) *EvmRegistry {
 		registry: mockRegistry,
 		abi:      keeperRegistryABI,
 		active:   make(map[string]activeUpkeep),
-		packer:   &evmRegistryPackerV2_1{abi: keeperRegistryABI},
+		packer:   NewEvmRegistryPackerV2_1(keeperRegistryABI),
 		headFunc: func(ocr2keepers.BlockKey) {},
 		chLog:    make(chan logpoller.Log, 1000),
 		mercury: &MercuryConfig{
